requirements.txt
sqlalchemy
pyTelegramBotAPI
python-decouple
requests
alembic
snowflake-sqlalchemy
passlib
pydantic
openai
python-dotenv
redis
celery
PyPDF2
python-docx
pydub
psycopg2-binary
pycryptodome
ocrmypdf
pytube
